Move donut 3d path components into own JS module
